Fix: cast from ‘char*’ to ‘unsigned int’ loses precision

+++ auth_tcpriv.cc
@@ -69,7 +69,7 @@ static int tcpriv_auth(MYSQL_PLUGIN_VIO *vio, MYSQL_SERVER_AUTH_INFO *info)
   MYSQL_PLUGIN_VIO_INFO vio_info;
   unsigned char *pkt;
   unsigned char syn[500];
-  unsigned int cli_uid;
+  uint64_t cli_uid;
   socklen_t syn_len = sizeof(syn);
   tcpriv_info tinfo;
 
@@ -79,7 +79,7 @@ static int tcpriv_auth(MYSQL_PLUGIN_VIO *vio, MYSQL_SERVER_AUTH_INFO *info)
   }
 
   //info->password_used = PASSWORD_USED_NO_MENTION;
-  cli_uid = (unsigned int)info->user_name;
+  cli_uid = (uint64_t)info->user_name;
 
   vio->info(vio, &vio_info);
   if (vio_info.protocol != MYSQL_PLUGIN_VIO_INFO::MYSQL_VIO_SOCKET)
